all: Remove IndexedRangeExpression
As Ali suggested, there is no need for an IndexedRangeExpression as we
can just use an IndexedExpression with the expression being a range.
